chore: fix compilation error
